Make sure the DB service could be generated. (#4757)
