+++ oap-server/server-receiver-plugin/skywalking-trace-receiver-plugin/src/main/java/org/apache/skywalking/oap/server/receiver/trace/provider/parser/listener/MultiScopesAnalysisListener.java
@@ -286,6 +286,7 @@ public class MultiScopesAnalysisListener implements EntryAnalysisListener, ExitA
                 sourceReceiver.receive(serviceInstanceRelation);
             }
             if (RequestType.DATABASE.equals(exitSourceBuilder.getType())) {
+                sourceReceiver.receive(exitSourceBuilder.toServiceMeta());
                 sourceReceiver.receive(exitSourceBuilder.toDatabaseAccess());
             }
         });
